feat(gui-v2): move grid view to nested page
